Hi Jovaugn! Your add button is broken.
It looks like the add button isn't working when I'm trying to be cheeky and add a new device.

I believe the issue is that you're starting with an input tag and ending with a button tag. Let me know if this works or not because I certainly haven't tested it.